mysql: widen BlockId to int, executed 2026-09-22
Blocks.BlockId and Lfns.BlockId were signed mediumint at 73.2% of their
8388607 ceiling.  Both are now int(11), ceiling 2147483647.

  ALTER TABLE Lfns   MODIFY BlockId int NOT NULL;
  ALTER TABLE Blocks MODIFY BlockId int NOT NULL AUTO_INCREMENT;

Cost 11s and 9s of table lock, so no outage window was needed -- a dry run
on a full copy of Lfns took 12s, which is what justified that.  Row counts,
max(BlockId) and the auto_increment counter are unchanged, and every Lfns row
still joins to a Block.  In-database backups Blocks_backup_20260922 and
Lfns_backup_20260922 were taken first and still exist.

Also: dump-schema.sh now skips scratch tables (_backup_, _test) so they do not
leak into schema.sql, and notes in a comment what it skipped.

The temporary ALTER/CREATE/DROP grants on ssluser still need revoking.

+++ mysql/schema.sql
@@ -11,6 +11,7 @@
 
 CREATE DATABASE IF NOT EXISTS Bambu;
 USE Bambu;
+-- Skipped as scratch (matching _backup_|_test$|_test_): Blocks_backup_20260922 Lfns_backup_20260922 
 
 CREATE TABLE `Datasets` (
   `DatasetId` mediumint(9) NOT NULL AUTO_INCREMENT,
@@ -25,7 +26,7 @@ CREATE TABLE `Datasets` (
 ) ENGINE=MyISAM DEFAULT CHARSET=latin1 COLLATE=latin1_swedish_ci;
 
 CREATE TABLE `Blocks` (
-  `BlockId` mediumint(9) NOT NULL AUTO_INCREMENT,
+  `BlockId` int(11) NOT NULL AUTO_INCREMENT,
   `DatasetId` mediumint(9) NOT NULL,
   `BlockName` char(36) NOT NULL,
   PRIMARY KEY (`BlockId`),
@@ -34,7 +35,7 @@ CREATE TABLE `Blocks` (
 
 CREATE TABLE `Lfns` (
   `DatasetId` mediumint(9) NOT NULL,
-  `BlockId` mediumint(9) NOT NULL,
+  `BlockId` int(11) NOT NULL,
   `FileName` varchar(255) DEFAULT NULL,
   `PathName` text DEFAULT NULL,
   `NEvents` mediumint(9) NOT NULL DEFAULT -1,
